# bonus/openapi/specs.openapi.yaml
openapi: 3.0.3
info:
  title: EpyTodo API
  version: 1.0.0
  description: RESTful API for a Todo application with user authentication
servers:
  - url: http://localhost:{port}
    variables:
      port:
        default: '3000'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: integer, default: 0 }
        email: { type: string, default: 'jean@exemple.com' }
        password: { type: string, default: '{{HASH_PASSWORD}}' }
        firstname: { type: string, default: 'Jean' }
        name: { type: string, default: 'Joe' }
        created_at: { type: string, default: '2021-03-03 19:24:00' }
...
    Message:
      type: object
      properties:
        msg: { type: string }
      required: [msg]
  responses:
    Unauthorized:
      description: Unauthorized - No token or invalid token
      content:
        application/json:
          examples:
            noToken:
              value: { msg: 'No token, authorization denied' }
            invalidToken:
              value: { msg: 'Token is not valid' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          example:
            msg: Not found
    BadRequest:
      description: Bad parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          example:
            msg: Bad parameter
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          example:
            msg: Internal server error
security:
  - BearerAuth: []
paths:
...
  /users/{identifier}:
    get:
      summary: Get a user by ID or email
      security: [{ BearerAuth: [] }]
      parameters:
        - name: identifier
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          {
            description: User data,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/User' } },
              },
          }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /users/{id}:
    put:
      summary: Update user information
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                email: { type: string }
                password: { type: string }
                firstname: { type: string }
                name: { type: string }
      responses:
        '204':
          {
            description: Updated user,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/User' } },
              },
          }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }

    delete:
      summary: Delete a user
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          {
            description: User deleted,
            content:
              {
                application/json:
                  {
                    schema: { $ref: '#/components/schemas/Message' },
                    example:
                      { msg: 'Successfully deleted record number: {id}' },
                  },
              },
          }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...